+++ modules/library/xml/src/main/java/org/geotools/GML.java
@@ -29,7 +29,6 @@ import java.util.Comparator;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Iterator;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -53,7 +52,6 @@ import org.eclipse.xsd.XSDTypeDefinition;
 import org.eclipse.xsd.util.XSDConstants;
 import org.eclipse.xsd.util.XSDResourceImpl;
 import org.geotools.data.DataUtilities;
-import org.geotools.data.collection.ListFeatureCollection;
 import org.geotools.data.simple.SimpleFeatureCollection;
 import org.geotools.data.simple.SimpleFeatureIterator;
 import org.geotools.data.store.ReTypingFeatureCollection;
@@ -69,6 +67,7 @@ import org.geotools.gml2.GMLConfiguration;
 import org.geotools.gml2.bindings.GML2ParsingUtils;
 import org.geotools.gtxml.GTXML;
 import org.geotools.referencing.CRS;
+import org.geotools.util.PartiallyOrderedSet;
 import org.geotools.xml.Configuration;
 import org.geotools.xml.Encoder;
 import org.geotools.xml.Parser;
